pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nase"
version = "0.1.0"
description = "Noise-Adaptive Spectral Embedding toolkit and experiments."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "NASE contributors" }]
dependencies = [
  "numpy",
  "scipy",
  "scikit-learn",
  "matplotlib",
  "pandas",
  "pyyaml",
  "typer",
  "click<8.2",
]

[project.optional-dependencies]
dev = [
  "pytest",
  "ruff",
  "pre-commit",
]

[project.scripts]
nase = "nase.cli:app"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "N", "W"]
ignore = ["E203"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-q"

